Add GitHub Actions workflow for Hugo site deployment
Created a new GitHub Actions workflow to automate the deployment of the Hugo site to GitHub Pages. The workflow includes steps for installing Hugo, Dart Sass, building the site, and deploying it to the specified GitHub Pages environment.

Updated the baseURL in the Hugo configuration to reflect the new domain (3tl.org) and added a CNAME file for custom domain support.